Change calculate to parseFloat instead of parseInt

+++ script.js
@@ -32,7 +32,7 @@ function calculate(e) {
       operator = e.target.textContent;
       screen.textContent = num1 + operator + num2;
     } else {
-      screen.textContent = operate(operator, parseInt(num1), parseInt(num2));
+      screen.textContent = operate(operator, parseFloat(num1), parseFloat(num2));
       operator = e.target.textContent;
       screen.textContent = num1 + operator + num2;
     }
